requirements.txt
Django==1.8.7
celery==3.1.17
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.10.0
django-grappelli==2.6.5
django-nose==1.4.1
django-rest-swagger==0.3.2
django-sslify==0.2.7
django-suit==0.2.13
django-url-filter==0.2.0
djangorestframework==3.1.3
djangorestframework-filters==0.3.0
fakeredis==0.6.1
gunicorn==19.0.0
httpretty==0.8.3
mock==1.0.1
jsonfield==1.0.3
nose==1.3.4
psycopg2==2.5.3
redis==2.10.3
requests==2.3.0
sendwithus==1.0.22
static3==0.5.1
urllib3==1.7.1
wsgiref==0.1.2
newrelic==2.40.0.34
oauthlib==0.7.2
django-model-utils==2.2
django-redis==4.2.0
django-haystack==2.4.0
elasticsearch>=1.0.0,<2.0.0
bleach==1.4.1
Markdown==2.6.2
django-notifications-hq>=0.7.1
django_select2==4.3.1
django-suit-ckeditor==0.0.2
feedparser==5.2.1

### Note: using our own forks for the listed libraries below:
djangorestframework-bulk==0.2.1
# git+git://github.com/purple-desk/django-rest-framework-bulk.git
# django-counter-field==0.3.1
git+git://github.com/purple-desk/django-counter-field.git
############################

### Please don't change this
rest_framework_authtoken_cookie-0.5.tar.gz
############################
